backend/requirements/base.txt
# Django core
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8

# Authentication
django-allauth==65.4.1
dj-rest-auth==7.0.0
djangorestframework-simplejwt==5.4.0
requests==2.32.5

# Database
psycopg2-binary==2.9.10
pgvector==0.3.6

# Task queue
celery==5.4.0
redis==5.2.1  # BSD-3-Clause Python client; server uses Valkey (BSD-3-Clause)

# Production server
gunicorn==23.0.0

# Content scraping
youtube-transcript-api==0.6.3  # MIT
beautifulsoup4==4.12.3  # MIT
trafilatura==1.12.2  # Apache-2.0
lxml==5.3.0  # BSD

# Text processing & embeddings
sentence-transformers==3.3.1  # Apache-2.0
tiktoken==0.8.0  # MIT

# LLM API
anthropic>=0.40.0  # MIT - Claude API client
openai>=1.0.0  # MIT - OpenAI API client (quiz generation)

# ASGI server for streaming
uvicorn[standard]>=0.32.0  # BSD-3-Clause
